cleanup: general cleanup of wip commits and restructuring translation folder
