+++ src/components/Spinner.tsx
@@ -11,7 +11,7 @@ const spinnerVariants = cva("text-muted-foreground animate-spin slower", {
       sm: "h-2 w-2",
       lg: "h-6 w-6",
       icon: "h-10 w-10",
-      screen: "h-20 w-20"
+      screen: "h-20 w-20",
     },
   },
   defaultVariants: {
@@ -22,5 +22,12 @@ const spinnerVariants = cva("text-muted-foreground animate-spin slower", {
 interface SpinnerProps extends VariantProps<typeof spinnerVariants> {}
 
 export const Spinner = ({ size }: SpinnerProps) => {
-  return <Loader className={cn("text-muted-foreground animate-spin slower", spinnerVariants({ size }))} />;
+  return (
+    <Loader
+      className={cn(
+        "text-muted-foreground animate-spin slower",
+        spinnerVariants({ size })
+      )}
+    />
+  );
 };
