+++ src/lib/translations/applicant-form/thirdform_es.json
